Persistence for Book class

+++ src/main/java/com/nielsen/convensiondemo/manager/BookManager.java
@@ -6,6 +6,6 @@ import com.nielsen.convensiondemo.model.Book;
 
 public interface BookManager {
 	List<Book> list();
-	Book get(Integer bookId);
+	Book get(Long bookId);
 	String save(Book book);
 }
